Updates with Poplar SDK 3.0 release

+++ utils/triton_server/server_setup.py
@@ -12,11 +12,10 @@ import tempfile
 from .server import TritonServer
 
 model_repo_opt = "--model-repository"
-backend_dir_opt = "--backend-directory"
 grpc_port_opt = "--grpc-port"
 benchmark_opt = "--benchmark_only"
 
-backend_dir_def = str(Path(__file__).parent.absolute()) + "/backends"
+triton_server_build_rel_path = "/server/mybuild/tritonserver/build/server/mybuild/tritonserver/install/bin/"
 
 
 def next_free_port(port=2000, max_port=65535):
@@ -32,10 +31,13 @@ def next_free_port(port=2000, max_port=65535):
 
 
 def pytest_addoption(parser):
-    parser.addoption(model_repo_opt, action="store", default=None)
-    parser.addoption(backend_dir_opt, action="store", default=backend_dir_def)
-    parser.addoption(grpc_port_opt, action="store", default=next_free_port())
-    parser.addoption(benchmark_opt, action="store", default=False)
+    parser.addoption(model_repo_opt, action="store", default=None,
+                     help="Path to directory which consists model configuration files for Triton Server \
+and model exported to popef file. Default: `TEST_ROOT_PATH/models`")
+    parser.addoption(grpc_port_opt, action="store", default=next_free_port(),
+                     help="Port on which Triton Server will be listening for commands.")
+    parser.addoption(benchmark_opt, action="store", default=False,
+                     help="Run benchmarks sending requests with exactly 1 batch size of selected model.")
 
 
 triton_env_lockfile_name = str(
@@ -50,23 +52,30 @@ def prepare_triton_environment(request):
     if not poplar_sdk_dir_env:
         pytest.fail("Environment variable POPLAR_SDK_ENABLED is not set!")
 
-    # create triton backend folder - copy files from SDK and arrange it in structure
-    # expected by Triton
-    poplar_backend_dir = str(
-        request.config.getoption(backend_dir_opt)) + "/poplar"
-    os.system('mkdir -p ' + poplar_backend_dir)
-    os.system('cp -rfL ' + poplar_sdk_dir_env + '/lib ' + poplar_backend_dir)
-    os.system('cp -fL ' + poplar_sdk_dir_env + '/../libtriton_poplar.so ' +
-              poplar_backend_dir + '/libtriton_poplar.so')
-
-    print("Compiling Triton Server.")
+    print("Prepare Triton Server.")
     try:
         ts_dir = tempfile.mkdtemp()
         with open(triton_env_lockfile_name, "w") as telfn:
-            telfn.write(str(ts_dir))
+            telfn.write(str(ts_dir) + triton_server_build_rel_path)
         subprocess.run(["sh", "download_and_build_triton_server.sh", str(ts_dir)],
                        cwd=str(Path(__file__).parent.absolute()), check=True,
                        stderr=subprocess.PIPE)
+        # create Triton backend folder - copy files from SDK and arrange it in structure
+        # expected by Triton
+        poplar_backend_dir = ts_dir + triton_server_build_rel_path + "/poplar"
+        os.system('mkdir -p ' + poplar_backend_dir)
+        os.system('cp -rfL ' + poplar_sdk_dir_env +
+                  '/lib ' + poplar_backend_dir)
+        if Path(poplar_sdk_dir_env + '/../libtriton_poplar.so').is_file():
+            os.system('cp -fL ' + poplar_sdk_dir_env + '/../libtriton_poplar.so ' +
+                      poplar_backend_dir + '/libtriton_poplar.so')
+        elif Path('/opt/triton/lib/libtriton_poplar.so').is_file():
+            os.system('cp -fL /opt/triton/lib/libtriton_poplar.so ' +
+                      poplar_backend_dir + '/libtriton_poplar.so')
+        else:
+            pytest.fail(
+                "Can't find library \"libtriton_poplar.so\" needed to run Triton Server.")
+
     except subprocess.CalledProcessError as e:
         pytest.fail(
             "Failed to download and/or compile Triton Server! stderr: " + str(e.stderr))
@@ -75,13 +84,11 @@ def prepare_triton_environment(request):
 @pytest.fixture(scope="module")
 def triton_server(request, prepare_triton_environment):
     with open(triton_env_lockfile_name, "r") as telfn:
-        ts_exe_path = telfn.read()
-    ts_exe_path = ts_exe_path + \
-        "/server/mybuild/tritonserver/build/server/mybuild/tritonserver/install/bin/tritonserver"
-    if not Path(ts_exe_path).is_file():
-        pytest.fail(str(ts_exe_path) + " doesnt exists!")
-    server = TritonServer(request, ts_exe_path, model_repo_opt,
-                          backend_dir_opt, grpc_port_opt)
+        ts_exec_path = telfn.read()
+    if not Path(ts_exec_path + "/tritonserver").is_file():
+        pytest.fail(str(ts_exec_path) + "/tritonserver doesnt exists!")
+    server = TritonServer(request, ts_exec_path, model_repo_opt,
+                          grpc_port_opt)
     yield server
     server.terminate()
